Remove PluggableMap
