+++ src/ol/MapBrowserEventHandler.js
@@ -13,7 +13,7 @@ import {listen, unlistenByKey} from './events.js';
 
 class MapBrowserEventHandler extends Target {
   /**
-   * @param {import("./PluggableMap.js").default} map The map with the viewport to listen to events on.
+   * @param {import("./Map.js").default} map The map with the viewport to listen to events on.
    * @param {number} [moveTolerance] The minimal distance the pointer must travel to trigger a move.
    */
   constructor(map, moveTolerance) {
@@ -21,7 +21,7 @@ class MapBrowserEventHandler extends Target {
 
     /**
      * This is the element that we will listen to the real events on.
-     * @type {import("./PluggableMap.js").default}
+     * @type {import("./Map.js").default}
      * @private
      */
     this.map_ = map;
