+++ src/ol/Overlay.js
@@ -229,13 +229,13 @@ class Overlay extends BaseObject {
 
   /**
    * Get the map associated with this overlay.
-   * @return {import("./PluggableMap.js").default|null} The map that the
+   * @return {import("./Map.js").default|null} The map that the
    * overlay is part of.
    * @observable
    * @api
    */
   getMap() {
-    return /** @type {import("./PluggableMap.js").default|null} */ (
+    return /** @type {import("./Map.js").default|null} */ (
       this.get(Property.MAP) || null
     );
   }
@@ -356,7 +356,7 @@ class Overlay extends BaseObject {
 
   /**
    * Set the map to be associated with this overlay.
-   * @param {import("./PluggableMap.js").default|null} map The map that the
+   * @param {import("./Map.js").default|null} map The map that the
    * overlay is part of. Pass `null` to just remove the overlay from the current map.
    * @observable
    * @api
