+++ src/ol/interaction/DragAndDrop.js
@@ -254,7 +254,7 @@ class DragAndDrop extends Interaction {
    * Remove the interaction from its current map and attach it to the new map.
    * Subclasses may set up event handlers to get notified about changes to
    * the map here.
-   * @param {import("../PluggableMap.js").default} map Map.
+   * @param {import("../Map.js").default} map Map.
    */
   setMap(map) {
     this.unregisterListeners_();
