+++ src/ol/interaction/Translate.js
@@ -336,7 +336,7 @@ class Translate extends PointerInteraction {
    * Tests to see if the given coordinates intersects any of our selected
    * features.
    * @param {import("../pixel.js").Pixel} pixel Pixel coordinate to test for intersection.
-   * @param {import("../PluggableMap.js").default} map Map to test the intersection on.
+   * @param {import("../Map.js").default} map Map to test the intersection on.
    * @return {import("../Feature.js").default} Returns the feature found at the specified pixel
    * coordinates.
    * @private
@@ -381,7 +381,7 @@ class Translate extends PointerInteraction {
    * Remove the interaction from its current map and attach it to the new map.
    * Subclasses may set up event handlers to get notified about changes to
    * the map here.
-   * @param {import("../PluggableMap.js").default} map Map.
+   * @param {import("../Map.js").default} map Map.
    */
   setMap(map) {
     const oldMap = this.getMap();
@@ -397,7 +397,7 @@ class Translate extends PointerInteraction {
   }
 
   /**
-   * @param {import("../PluggableMap.js").default} oldMap Old map.
+   * @param {import("../Map.js").default} oldMap Old map.
    * @private
    */
   updateState_(oldMap) {
