+++ src/ol/layer/MapboxVector.js
@@ -88,10 +88,10 @@ class ErrorEvent extends BaseEvent {
  *  * `'vector'`: Everything is rendered as vectors. Use this mode for improved performance on vector
  *    tile layers with only a few rendered features (e.g. for highlighting a subset of features of
  *    another layer with the same source).
- * @property {import("../PluggableMap.js").default} [map] Sets the layer as overlay on a map. The map will not manage
+ * @property {import("../Map.js").default} [map] Sets the layer as overlay on a map. The map will not manage
  * this layer in its layers collection, and the layer will be rendered on top. This is useful for
  * temporary layers. The standard way to add a layer to a map and have it managed by the map is to
- * use {@link import("../PluggableMap.js").default#addLayer map.addLayer()}.
+ * use [map.addLayer()]{@link import("../Map.js").default#addLayer}.
  * @property {boolean} [updateWhileAnimating=false] When set to `true`, feature batches will be
  * recreated during animations. This means that no vectors will be shown clipped, but the setting
  * will have a performance impact for large amounts of vector data. When set to `false`, batches
