+++ src/ol/renderer/Layer.js
@@ -58,7 +58,7 @@ class LayerRenderer extends Observable {
   /**
    * Determine whether render should be called.
    * @abstract
-   * @param {import("../PluggableMap.js").FrameState} frameState Frame state.
+   * @param {import("../Map.js").FrameState} frameState Frame state.
    * @return {boolean} Layer is ready to be rendered.
    */
   prepareFrame(frameState) {
@@ -68,7 +68,7 @@ class LayerRenderer extends Observable {
   /**
    * Render the layer.
    * @abstract
-   * @param {import("../PluggableMap.js").FrameState} frameState Frame state.
+   * @param {import("../Map.js").FrameState} frameState Frame state.
    * @param {HTMLElement} target Target that may be used to render content to.
    * @return {HTMLElement} The rendered element.
    */
@@ -116,7 +116,7 @@ class LayerRenderer extends Observable {
   /**
    * @abstract
    * @param {import("../coordinate.js").Coordinate} coordinate Coordinate.
-   * @param {import("../PluggableMap.js").FrameState} frameState Frame state.
+   * @param {import("../Map.js").FrameState} frameState Frame state.
    * @param {number} hitTolerance Hit tolerance in pixels.
    * @param {import("./vector.js").FeatureCallback<T>} callback Feature callback.
    * @param {Array<import("./Map.js").HitMatch<T>>} matches The hit detected matches with tolerance.
