+++ src/ol/renderer/Map.js
@@ -25,14 +25,14 @@ import {wrapX} from '../coordinate.js';
  */
 class MapRenderer extends Disposable {
   /**
-   * @param {import("../PluggableMap.js").default} map Map.
+   * @param {import("../Map.js").default} map Map.
    */
   constructor(map) {
     super();
 
     /**
      * @private
-     * @type {import("../PluggableMap.js").default}
+     * @type {import("../Map.js").default}
      */
     this.map_ = map;
   }
@@ -40,14 +40,14 @@ class MapRenderer extends Disposable {
   /**
    * @abstract
    * @param {import("../render/EventType.js").default} type Event type.
-   * @param {import("../PluggableMap.js").FrameState} frameState Frame state.
+   * @param {import("../Map.js").FrameState} frameState Frame state.
    */
   dispatchRenderEvent(type, frameState) {
     abstract();
   }
 
   /**
-   * @param {import("../PluggableMap.js").FrameState} frameState FrameState.
+   * @param {import("../Map.js").FrameState} frameState FrameState.
    * @protected
    */
   calculateMatrices2D(frameState) {
@@ -71,7 +71,7 @@ class MapRenderer extends Disposable {
 
   /**
    * @param {import("../coordinate.js").Coordinate} coordinate Coordinate.
-   * @param {import("../PluggableMap.js").FrameState} frameState FrameState.
+   * @param {import("../Map.js").FrameState} frameState FrameState.
    * @param {number} hitTolerance Hit tolerance in pixels.
    * @param {boolean} checkWrapped Check for wrapped geometries.
    * @param {import("./vector.js").FeatureCallback<T>} callback Feature callback.
@@ -172,7 +172,7 @@ class MapRenderer extends Disposable {
 
   /**
    * @param {import("../coordinate.js").Coordinate} coordinate Coordinate.
-   * @param {import("../PluggableMap.js").FrameState} frameState FrameState.
+   * @param {import("../Map.js").FrameState} frameState FrameState.
    * @param {number} hitTolerance Hit tolerance in pixels.
    * @param {boolean} checkWrapped Check for wrapped geometries.
    * @param {function(this: U, import("../layer/Layer.js").default): boolean} layerFilter Layer filter
@@ -206,7 +206,7 @@ class MapRenderer extends Disposable {
   }
 
   /**
-   * @return {import("../PluggableMap.js").default} Map.
+   * @return {import("../Map.js").default} Map.
    */
   getMap() {
     return this.map_;
@@ -215,14 +215,14 @@ class MapRenderer extends Disposable {
   /**
    * Render.
    * @abstract
-   * @param {?import("../PluggableMap.js").FrameState} frameState Frame state.
+   * @param {?import("../Map.js").FrameState} frameState Frame state.
    */
   renderFrame(frameState) {
     abstract();
   }
 
   /**
-   * @param {import("../PluggableMap.js").FrameState} frameState Frame state.
+   * @param {import("../Map.js").FrameState} frameState Frame state.
    * @protected
    */
   scheduleExpireIconCache(frameState) {
@@ -233,8 +233,8 @@ class MapRenderer extends Disposable {
 }
 
 /**
- * @param {import("../PluggableMap.js").default} map Map.
- * @param {import("../PluggableMap.js").FrameState} frameState Frame state.
+ * @param {import("../Map.js").default} map Map.
+ * @param {import("../Map.js").FrameState} frameState Frame state.
  */
 function expireIconCache(map, frameState) {
   iconImageCache.expire();
