+++ src/ol/source/Source.js
@@ -11,10 +11,10 @@ import {get as getProjection} from '../proj.js';
  */
 
 /**
- * A function that takes a {@link module:ol/PluggableMap~FrameState} and returns a string or
+ * A function that takes a {@link module:ol/Map~FrameState} and returns a string or
  * an array of strings representing source attributions.
  *
- * @typedef {function(import("../PluggableMap.js").FrameState): (string|Array<string>)} Attribution
+ * @typedef {function(import("../Map.js").FrameState): (string|Array<string>)} Attribution
  */
 
 /**
