+++ src/ol/webgl/PostProcessingPass.js
@@ -191,7 +191,7 @@ class WebGLPostProcessingPass {
    * Initialize the render target texture of the post process, make sure it is at the
    * right size and bind it as a render target for the next draw calls.
    * The last step to be initialized will be the one where the primitives are rendered.
-   * @param {import("../PluggableMap.js").FrameState} frameState current frame state
+   * @param {import("../Map.js").FrameState} frameState current frame state
    * @api
    */
   init(frameState) {
@@ -250,10 +250,10 @@ class WebGLPostProcessingPass {
 
   /**
    * Render to the next postprocessing pass (or to the canvas if final pass).
-   * @param {import("../PluggableMap.js").FrameState} frameState current frame state
+   * @param {import("../Map.js").FrameState} frameState current frame state
    * @param {WebGLPostProcessingPass} [nextPass] Next pass, optional
-   * @param {function(WebGLRenderingContext, import("../PluggableMap.js").FrameState):void} [preCompose] Called before composing.
-   * @param {function(WebGLRenderingContext, import("../PluggableMap.js").FrameState):void} [postCompose] Called before composing.
+   * @param {function(WebGLRenderingContext, import("../Map.js").FrameState):void} [preCompose] Called before composing.
+   * @param {function(WebGLRenderingContext, import("../Map.js").FrameState):void} [postCompose] Called before composing.
    * @api
    */
   apply(frameState, nextPass, preCompose, postCompose) {
@@ -325,7 +325,7 @@ class WebGLPostProcessingPass {
 
   /**
    * Sets the custom uniforms based on what was given in the constructor.
-   * @param {import("../PluggableMap.js").FrameState} frameState Frame state.
+   * @param {import("../Map.js").FrameState} frameState Frame state.
    * @private
    */
   applyUniforms(frameState) {
